Tests/LibWeb/Text/expected/WebAnimations/misc/parse-animation-timing-function-steps.txt

Computed animation-timing-function: ease
